requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==4.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
docopt==0.6.2
elevenlabs==2.8.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Js2Py==0.74
lxml==5.4.0
MouseInfo==0.1.3
numpy==2.3.0
packaging==25.0
pipwin==0.5.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
pyperclip==1.9.0
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-docx==1.1.2
python-dotenv==1.1.1
python-telegram-bot==13.15
pytweening==1.2.0
pytz==2025.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
rubicon-objc==0.5.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
spotipy==2.25.1
standard-aifc==3.13.0
standard-chunk==3.13.0
tornado==6.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.16
websockets==15.0.1
